EE: Move EE code to new files
